depreciated/requirements.txt
pyyaml
networkx
numpy
matplotlib
pandas
